mv safe_tuple, force_1d_inputs from complexity to utils
